feat(toggle-theme): add new ToggleTheme component for theme switching
Add a new ToggleTheme component to allow users to switch between
light and dark themes using the next-themes library. This enhances
user experience by providing a customizable interface.

feat(use-is-client): introduce useIsClient hook for client-side checks

Add a useIsClient hook to determine if the code is running on the
client or server side. This is essential for components that rely
on client-specific features.

chore(package.json): update version to 2025.2.6

Update the package version to reflect the addition of new features
and components.

refactor(header): replace SwitchTheme with ToggleTheme component

Update the Header component to use the new ToggleTheme component
instead of the deprecated SwitchTheme, ensuring consistency with
the new theme switching implementation.